refactor: 예외 처리 분리
