+++ src/main/java/racingcar/controller/exception/ExceptionResponse.java
@@ -1,0 +1,14 @@
+package racingcar.controller.exception;
+
+public class ExceptionResponse {
+
+    private final String exceptionMessage;
+
+    public ExceptionResponse(final String exceptionMessage) {
+        this.exceptionMessage = exceptionMessage;
+    }
+
+    public String getExceptionMessage() {
+        return exceptionMessage;
+    }
+}
